+++ errors_utils/db_errors.go
@@ -1,4 +1,4 @@
-package errors
+package errors_utils
 
 import (
 	"database/sql"
