+++ errors_utils/errors.go
@@ -1,4 +1,4 @@
-package errors
+package errors_utils
 
 import (
 	"fmt"
