+++ errors_utils/messages.go
@@ -1,3 +1,3 @@
-package errors
+package errors_utils
 
 const ErrInvalidPassword = "Пароль должен содержать только латинские буквы, цифры и спец символы. Должны быть минимум 1 строчная, 1 заглавная, 1 цифра и 1 один спецсимвол"
