+++ utils/utils.go
@@ -2,11 +2,11 @@ package utils
 
 import (
 	"github.com/gofiber/fiber/v2"
-	"github.com/rogerok/wflow-backend/errors"
+	"github.com/rogerok/wflow-backend/errors_utils"
 	"github.com/rogerok/wflow-backend/responses"
 )
 
-func GetResponseError(ctx *fiber.Ctx, err *errors.CustomError) error {
+func GetResponseError(ctx *fiber.Ctx, err *errors_utils.CustomError) error {
 	return ctx.Status(err.StatusCode).JSON(err)
 }
